==> analysis/01_prep-decennial-resources.qmd <==
---
title: "Setup and prep"
format: 
  html:
    embed-resources: true
    toc: true
editor: source
execute:
  echo: true
  warning: false
  message: false
---

## Setup and prep

This document contains setup/prep steps for the opt-out privacy workflow.

```{r setup}
library(tidyverse)
library(readxl)
library(here)

options(scipen = 999)
```

### Prep Stateside PUMS fixed-width layout file

The excel layout file provides variable widths, names, dictionary values, etc. for the Census Stateside PUMS microdata. Use this layout file to help process the fixed-width stateside PUMS data.

```{r}
person_layout <- read_excel(here("data", "resources", "2010 PUMS Record Layout.xlsx"), sheet = "Person", skip = 1)
saveRDS(layout, file = paste(here("data", "layout"), "fwf-person-layout.rds", sep = "/"))

housing_layout <- read_excel(here("data", "resources", "2010 PUMS Record Layout.xlsx"), sheet = "Housing", skip = 1)
saveRDS(layout, file = paste(here("data", "layout"), "fwf-housing-layout.rds", sep = "/"))

layout <- rbind(person_layout, housing_layout)
saveRDS(layout, file = paste(here("data", "layout"), "fwf-layout.rds", sep = "/"))
```

### Prep value description fields

Variable values are encoded. Prep layout file dictionary description ("_VAL") fields to merge with data for easier use.

```{r}
layout_dict <- layout %>%
  filter(is.na(HI) & !is.na(LO) & VARIABLE != "RECTYPE") %>%
  select(c(VARIABLE, `VALUE DESCRIPTION`, LO)) %>%
  group_by(VARIABLE) %>%
  filter(n() > 1) %>%
  ungroup()
saveRDS(layout_dict, file = paste(here("data", "layout"), "fwf-layout-dict.rds", sep = "/"))
```

### Prep layout dictionary and helper list

Create a list of variables. 

```{r}
layout_dictlist <- unique(layout_dict$VARIABLE)
```

Pull values for different variables and save them in a layout file.

```{r}
prep_layout_dict <- function(variable) {
  
  helper_dict <- layout_dict %>%
    filter(VARIABLE == variable) %>%
    select(-c(VARIABLE)) %>%
    rename(!!as.name(variable) := LO,
           !!paste0(variable, "_VAL") := `VALUE DESCRIPTION`)
  
  return(helper_dict)
}
```

```{r}
layout_list <- map(layout_dictlist, prep_layout_dict)
saveRDS(layout_list, file = paste(here("data", "layout"), "fwf-layout-list.rds", sep = "/"))
```
